+++ demos/tutorial/AMR_1D_Burgers/step_4/update_field.hpp
@@ -48,7 +48,7 @@ void update_field(Field& f, const Tag& tag, Mesh& new_mesh)
     samurai::for_each_interval(mesh[mesh_id_t::cells],
                                [&](std::size_t level, const auto& interval, const auto&)
                                {
-                                   auto itag = static_cast<std::size_t>(interval.start + interval.index);
+                                   auto itag = interval.start + interval.index;
                                    for (coord_index_t i = interval.start; i < interval.end; ++i)
                                    {
                                        if (tag[itag] & static_cast<int>(samurai::CellFlag::refine))
