+++ demos/tutorial/AMR_1D_Burgers/step_4/update_mesh.hpp
@@ -41,7 +41,7 @@ bool update_mesh(Field& f, const Tag& tag)
     samurai::for_each_interval(mesh[mesh_id_t::cells],
                                [&](std::size_t level, const auto& interval, const auto&)
                                {
-                                   auto itag = static_cast<std::size_t>(interval.start + interval.index);
+                                   auto itag = interval.start + interval.index;
                                    for (int i = interval.start; i < interval.end; ++i)
                                    {
                                        if (tag[itag] & static_cast<int>(samurai::CellFlag::refine))
