+++ include/samurai/algorithm/update.hpp
@@ -489,7 +489,7 @@ namespace samurai
         for_each_interval(mesh[mesh_id_t::cells],
                           [&](std::size_t level, const auto& interval, const auto& index)
                           {
-                              auto itag = static_cast<std::size_t>(interval.start + interval.index);
+                              auto itag = interval.start + interval.index;
                               for (auto i = interval.start; i < interval.end; ++i)
                               {
                                   if (tag[itag] & static_cast<int>(CellFlag::refine))
@@ -555,7 +555,7 @@ namespace samurai
         for_each_interval(mesh[mesh_id_t::cells],
                           [&](std::size_t level, const auto& interval, const auto& index)
                           {
-                              auto itag = static_cast<std::size_t>(interval.start + interval.index);
+                              auto itag = interval.start + interval.index;
                               for (value_t i = interval.start; i < interval.end; ++i)
                               {
                                   if (tag[itag] & static_cast<int>(CellFlag::refine))
